#4899 Recover the orphaned MobileObjectMetastateTests
Source/Csla.test/Serialization/MobileObjectMetastateTests.cs sat outside
Source/tests, in a directory with no project file, so nothing compiled it. It
was stranded by #4712 and edited twice since, in #4750 and #4785, meaning nine
metastate tests never ran. Moving it into the live project produced 12 compiler
errors: the file was not stale, it was never valid.

The BrokenRule tests move to csla.netcore.test, the assembly Csla grants
InternalsVisibleTo, since BrokenRule's populating ctor is internal. Its
parameterless ctor is obsolete-as-error, which no pragma can suppress, so the
empty instance is created by reflection exactly as MobileFormatter does.

The CommandBase test used an unrelated test class as an ObjectFactory and called
the protected LoadProperty on it; a small test-local ObjectFactory subclass now
exposes that. The flag preservation test called the protected MarkOld(); it now
fetches an object through the data portal, which is the state it was simulating.

Four of the nine then fail, all asserting that property values survive a
metastate round trip. That is what IMobileObjectMetastate's documentation
promises, but every implementation writes only state flags and undo bookkeeping,
and BrokenRule and CommandBase do not override OnGetMetastate at all. Whether
the docs overpromise or the implementations are incomplete is a maintainer
decision, so those four land [Ignore]d pointing at #4898 rather than being
deleted or bent to pass.

Five tests now run that never have.
